Refactor config loading and bundle data classes
Introduces SboConfigBundle to group SboData and AchievementsData, refactors config loading to use loadAllData, and updates references throughout the codebase. Also renames Data.kt to Configs.kt and adds AchievementsData data class.
