+++ src/main/kotlin/net/sbo/mod/SBOKotlin.kt
@@ -10,17 +10,16 @@ import net.sbo.mod.settings.Settings
 import net.sbo.mod.utils.Register
 import net.sbo.mod.utils.World
 import net.sbo.mod.general.PartyCommands
+import net.sbo.mod.data.SboConfigBundle
 import net.sbo.mod.data.SboDataObject
-import net.sbo.mod.data.configs.SboData
 import net.sbo.mod.utils.SboKeyBinds
-import net.sbo.mod.utils.Chat.chat
 
 object SBOKotlin {
 	@JvmField
 	val mc: MinecraftClient = MinecraftClient.getInstance()
 	private const val MOD_ID = "sbo-kotlin"
 	internal val logger = LoggerFactory.getLogger(MOD_ID)
-	lateinit var sboData: SboData
+	lateinit var SBOConfigBundle: SboConfigBundle
 
 	val configurator = Configurator("sbo")
 	val settings = Settings.register(configurator)
@@ -28,10 +27,10 @@ object SBOKotlin {
 	@JvmStatic
 	fun onInitializeClient() {
 		logger.info("Initializing SBO-Kotlin...")
+		SBOConfigBundle = SboDataObject.loadAllData("SBO")
 		registerHelpCommand()
 		WaypointManager
 		PartyCommands.registerPartyChatListeners()
-		sboData = SboDataObject.load("SBO", "SboData.json", SboData(), SboData::class.java)
 		Register.command("sbo") {
 			mc.send{
 				mc.setScreen(ResourcefulConfigScreen.getFactory("sbo").apply(null))
@@ -39,8 +38,9 @@ object SBOKotlin {
 		}
 		Register.command("sboTest") {
 			logger.info(World.isInSkyblock().toString())
-			logger.info(sboData.b2bInq.toString())
-
+			logger.info(SBOConfigBundle.sboData.b2bInq.toString())
+			logger.info(SBOConfigBundle.sboData.toString())
+			logger.info(SBOConfigBundle.achievementsData.toString())
 		}
 
 		SboKeyBinds.register()
